CMake: adopt GNUInstallDirs
This makes it a lot easier for packagers (it happens by default
if we've setup our environment correctly) to have files installed in the right
places, including libdir, but also documentation, which can vary.

This works cross-platform still but it means libgeotiff uses the
standard options which CMake projects tend to use for such locations
and therefore no special treatment is needed when packaging it.

Fixes: https://github.com/OSGeo/libgeotiff/issues/33
Signed-off-by: Sam James <sam@gentoo.org>